+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add `Expression.spread` for the spread operator `...`.
 * Add support 'late' field modifier.
 * Add support for `Expression.nullChecked` to add a null assertion operator.
+* Add support for creating `mixin`s.
 
 ## 4.0.0